move parsing of gpkg header to sqlite driver
