+++ geodiff/src/drivers/sqliteutils.cpp
@@ -493,3 +493,36 @@ std::vector<std::string> sqliteColumnNames(
 
   return az;
 }
+
+int parseGpkgbHeaderSize( const std::string &gpkgWkb )
+{
+  // see GPKG binary header definition http://www.geopackage.org/spec/#gpb_spec
+
+  char flagByte = gpkgWkb[ GPKG_FLAG_BYTE_POS ];
+
+  char envelope_byte = ( flagByte & GPKG_ENVELOPE_SIZE_MASK ) >> 1;
+  int envelope_size = 0;
+
+  switch ( envelope_byte )
+  {
+    case 1:
+      envelope_size = 32;
+      break;
+
+    case 2:
+    // fall through
+    case 3:
+      envelope_size = 48;
+      break;
+
+    case 4:
+      envelope_size = 64;
+      break;
+
+    default:
+      envelope_size = 0;
+      break;
+  }
+
+  return GPKG_NO_ENVELOPE_HEADER_SIZE + envelope_size;
+}
